convert ChangelogServiceTest to integration
